requirements.txt
streamlit
pytesseract
transformers
torch
Pillow
opencv-python-headless
requests
